feat:Add Course Manager Home and Payment

+++ learnhub-api/src/main/java/com/learnhub/common/TeacherDetailsResponse.java
@@ -1,6 +1,7 @@
 package com.learnhub.common;
 
 import java.util.List;
+
 import com.learnhub.course.Category;
 import com.learnhub.course.Course;
 import com.learnhub.user.teacher.Teacher;
@@ -17,12 +18,13 @@ public record TeacherDetailsResponse(
         String city,
         List<CourseListResponse> courses
 ) {
-    public record CourseListResponse(Long id, String name, CategoryResponse category, Double price) {
+    public record CourseListResponse(Long id, String name, CategoryResponse category, Float price) {
         public record CategoryResponse(Long id, String name) {
             public static CategoryResponse from(Category category) {
                 return new CategoryResponse(category.getId(), category.getName());
             }
         }
+
         public static CourseListResponse from(Course course) {
             return new CourseListResponse(
                     course.getId(),
